LTD-3455: split out tests
