+++ unit_tests/caseworker/queues/views/test_cases.py
@@ -1,5 +1,4 @@
 import pytest
-import uuid
 import re
 
 from bs4 import BeautifulSoup
@@ -7,18 +6,12 @@ from urllib import parse
 
 from django.urls import reverse
 
-import os
-
 from core import client
 from core.exceptions import ServiceError
 
 from caseworker.cases.helpers.case import LU_POST_CIRC_FINALISE_QUEUE_ALIAS, LU_PRE_CIRC_REVIEW_QUEUE_ALIAS
-from caseworker.queues.forms import CaseAssignmentUsersForm, CaseAssignmentQueueForm
-from caseworker.queues.views import CaseAssignmentsCaseAssigneeSteps
 
 queue_pk = "59ef49f4-cf0c-4085-87b1-9ac6817b4ba6"
-
-gov_user_id = "1f288b81-2c26-439f-ac32-2a43c8b1a5cb"
 
 default_params = {
     "page": ["1"],
@@ -26,8 +19,6 @@ default_params = {
     "selected_tab": ["all_cases"],
     "hidden": ["true"],
 }
-
-example_return_to_url = "/example/endpoint"
 
 
 @pytest.fixture(autouse=True)
@@ -52,7 +43,7 @@ def data_cases_search(mock_case_statuses, data_case_types, gov_uk_user_id):
         "results": {
             "cases": [
                 {
-                    "id": "094eed9a-23cc-478a-92ad-9a05ac17fad0",
+                    "id": "094eed9a-23cc-478a-92ad-9a05ac17fad0",  # /PS-IGNORE
                     "case_type": {
                         "id": "00000000-0000-0000-0000-000000000001",
                         "reference": {"key": "oiel", "value": "Open Individual Export Licence"},
@@ -85,7 +76,7 @@ def data_cases_search(mock_case_statuses, data_case_types, gov_uk_user_id):
                     "organisation": {},
                     "activity_updates": [
                         {
-                            "id": "02cc3048-f893-4f0a-b37f-d066bc0b072a",
+                            "id": "02cc3048-f893-4f0a-b37f-d066bc0b072a",  # /PS-IGNORE
                             "created_at": "2023-02-02T17:30:05.184293Z",
                             "user": {
                                 "id": "00000000-0000-0000-0000-000000000001",
@@ -98,10 +89,10 @@ def data_cases_search(mock_case_statuses, data_case_types, gov_uk_user_id):
                             "additional_text": "additional line1\r\nadditional line2\r\nadditional line3\r\nadditional line4\r\nadditional line5",
                         },
                         {
-                            "id": "77d3c3d4-9761-403a-9942-a2fcc41aa35d",
+                            "id": "77d3c3d4-9761-403a-9942-a2fcc41aa35d",  # /PS-IGNORE
                             "created_at": "2023-02-02T17:30:04.174597Z",
                             "user": {
-                                "id": "2eb6e0fa-5a5b-4db1-96cc-dd1473e0c636",
+                                "id": "2eb6e0fa-5a5b-4db1-96cc-dd1473e0c636",  # /PS-IGNORE
                                 "first_name": "Joe",
                                 "last_name": "Bloggs",
                                 "type": "exporter",
@@ -112,16 +103,16 @@ def data_cases_search(mock_case_statuses, data_case_types, gov_uk_user_id):
                         },
                     ],
                     "assignments": {
-                        "9c4e66be-9f0f-451a-9c5f-d30e9c4bb69d": {
-                            "email": "test@mail.com",
+                        "9c4e66be-9f0f-451a-9c5f-d30e9c4bb69d": {  # /PS-IGNORE
+                            "email": "test@mail.com",  # /PS-IGNORE
                             "first_name": "John",
                             "last_name": "Smith",
                             "queues": [{"id": "ee1a3870-73d7-4af3-b629-e28f2c2227d7", "name": "Initial Queue"}],
                             "team_id": "00000000-0000-0000-0000-000000000001",
                             "team_name": "Admin",
                         },
-                        "9c4e66be-9f0f-451a-9c5f-d30e9c4bb69e": {
-                            "email": "test2@mail.com",
+                        "9c4e66be-9f0f-451a-9c5f-d30e9c4bb69e": {  # /PS-IGNORE
+                            "email": "test2@mail.com",  # /PS-IGNORE
                             "first_name": "Joe",
                             "last_name": "Smith",
                             "queues": [{"id": "ee1a3870-73d7-4af3-b629-e28f2c2227d7", "name": "Initial Queue"}],
@@ -165,7 +156,7 @@ def data_cases_search(mock_case_statuses, data_case_types, gov_uk_user_id):
                     "submitted_at": "2023-01-16T14:53:09.826340Z",
                 },
                 {
-                    "id": "8fb76bed-fd45-4293-95b8-eda9468aa254",
+                    "id": "8fb76bed-fd45-4293-95b8-eda9468aa254",  # /PS-IGNORE
                     "case_type": {
                         "id": "00000000-0000-0000-0000-000000000004",
                         "reference": {"key": "siel", "value": "Standard Individual Export Licence"},
@@ -259,48 +250,6 @@ def mock_cases_search_error(requests_mock, queue_pk):
     return requests_mock.get(url=url, status_code=500, json={})
 
 
-def test_queues_cannot_be_created_and_modified(authorized_client, reset_config_users_list):
-    url = reverse("queues:manage")
-    response = authorized_client.get(url)
-    assert response.status_code == 200
-    assert response.context["can_change_config"] == False
-
-
-def test_queues_can_be_created_and_modified(authorized_client, specify_config_users_list):
-    url = reverse("queues:manage")
-    response = authorized_client.get(url)
-    assert response.status_code == 200
-    assert response.context["can_change_config"] == True
-
-
-@pytest.mark.parametrize(
-    "url",
-    (
-        (reverse("queues:add")),
-        (reverse("queues:edit", kwargs={"pk": "e9f8711e-b383-47e5-b160-153f27771234"})),
-    ),
-)
-def test_queues_add_view_returns_unauthorized_user_not_on_config_admin_list(
-    authorized_client, reset_config_users_list, url
-):
-    response = authorized_client.get(url)
-    assert response.status_code == 403
-    assert response.context["title"] == "Sorry, unauthorized"
-    assert response.context["description"] == "You don't have authorisation to view this page"
-
-
-@pytest.mark.parametrize(
-    "url",
-    (
-        (reverse("queues:add")),
-        (reverse("queues:edit", kwargs={"pk": "e9f8711e-b383-47e5-b160-153f27771234"})),
-    ),
-)
-def test_queues_add_view_returns_ok_user_on_config_admin_list(authorized_client, specify_config_users_list, url):
-    response = authorized_client.get(url)
-    assert response.status_code == 200
-
-
 @pytest.mark.parametrize(
     "url",
     [
@@ -312,12 +261,6 @@ def test_queues_add_view_returns_ok_user_on_config_admin_list(authorized_client,
 def test_cases_view(url, authorized_client):
     response = authorized_client.get(url)
     assert response.status_code == 200
-
-
-@pytest.fixture
-def mock_enforcement_xml_upload(requests_mock):
-    url = client._build_absolute_uri(f"/cases/enforcement-check/{queue_pk}/")
-    yield requests_mock.post(url=url)
 
 
 @pytest.fixture
@@ -348,39 +291,6 @@ def mock_team_cases_search(requests_mock, data_cases_search):
 def mock_cases_search_head(requests_mock):
     url = client._build_absolute_uri(f"/cases/")
     yield requests_mock.head(url=re.compile(f"{url}.*"), headers={"resource-count": "350"})
-
-
-@pytest.fixture
-def mock_enforcement_xml_validation_error(requests_mock):
-    url = client._build_absolute_uri(f"/cases/enforcement-check/{queue_pk}/")
-    data = {"errors": {"file": ["Invalid XML format received"]}}
-    yield requests_mock.post(url=url, status_code=400, json=data)
-
-
-def test_upload_enforcement_xml_valid_file(authorized_client, mock_enforcement_xml_upload, settings):
-    url = reverse("queues:enforcement_xml_import", kwargs={"pk": queue_pk})
-
-    file_path = os.path.join(settings.BASE_DIR, "unit_tests/caseworker/queues/example.xml")
-    data = {"file": open(file_path, "rb")}
-
-    response = authorized_client.post(url, data, format="multipart")
-
-    assert response.status_code == 302
-
-    with open(file_path, "r") as f:
-        assert mock_enforcement_xml_upload.last_request.json() == {"file": f.read()}
-
-
-def test_upload_enforcement_xml_invalid_file(authorized_client, mock_enforcement_xml_validation_error, settings):
-    url = reverse("queues:enforcement_xml_import", kwargs={"pk": queue_pk})
-
-    file_path = os.path.join(settings.BASE_DIR, "unit_tests/caseworker/queues/example.xml")
-    data = {"file": open(file_path, "rb")}
-
-    response = authorized_client.post(url, data, format="multipart")
-
-    assert response.status_code == 200
-    assert response.context_data["form"].errors == {"file": ["Invalid XML format received"]}
 
 
 def test_cases_home_page_view_context(authorized_client):
@@ -458,129 +368,6 @@ def test_trigger_list_checkbox_visible_checked(authorized_client):
     html = BeautifulSoup(response.content, "html.parser")
     checkbox = html.find(id="is-trigger-list")
     assert "checked" in checkbox.attrs
-
-
-def test_case_assignment_case_office_no_user_selected(authorized_client, mock_gov_users):
-    url = (
-        reverse("queues:case_assignments_case_officer", kwargs={"pk": queue_pk})
-        + f"?cases={str(uuid.uuid4())}&cases={str(uuid.uuid4())}"
-    )
-    data = {}
-    response = authorized_client.post(url, data)
-
-    assert response.status_code == 200
-    assert response.context_data["form"].errors == {
-        "users": ["Select a user to allocate as Licensing Unit case officer"]
-    }
-
-
-@pytest.fixture
-def post_to_step_user_assignment_with_return_to(post_to_step_factory, user_assignment_url):
-    user_assignment_url = user_assignment_url + f"&return_to={example_return_to_url}"
-    return post_to_step_factory(user_assignment_url)
-
-
-@pytest.fixture
-def post_to_step_user_assignment_with_invalid_return_to(post_to_step_factory, user_assignment_url):
-    user_assignment_url = user_assignment_url + "&return_to=http://www.evil.com"
-    return post_to_step_factory(user_assignment_url)
-
-
-def test_case_assignment_case_adviser_with_return_to(
-    mock_gov_users,
-    mock_team_queue,
-    mock_put_assignments,
-    post_to_step_user_assignment_with_return_to,
-):
-    data = {"users": [gov_user_id]}
-    response = post_to_step_user_assignment_with_return_to(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_USERS,
-        data,
-    )
-    assert response.status_code == 302
-    assert response.url == example_return_to_url
-
-
-def test_case_assignment_case_adviser_with_invalid_return_to(
-    post_to_step_user_assignment_with_invalid_return_to,
-):
-    data = {"users": [gov_user_id]}
-    response = post_to_step_user_assignment_with_invalid_return_to(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_USERS,
-        data,
-    )
-    assert response.status_code == 403
-    assert b"Invalid return_to parameter" in response.content
-
-
-@pytest.fixture
-def mock_put_case_case_officer(requests_mock):
-    case_officer_put_url = client._build_absolute_uri("/cases/cases-update-case-officer/")
-    return requests_mock.put(url=case_officer_put_url, json={})
-
-
-def test_case_assignment_case_officer(authorized_client, requests_mock, mock_gov_users, mock_put_case_case_officer):
-    cases_ids = [str(uuid.uuid4()), str(uuid.uuid4())]
-    url = (
-        reverse("queues:case_assignments_case_officer", kwargs={"pk": queue_pk})
-        + f"?cases={cases_ids[0]}&cases={cases_ids[1]}"
-    )
-
-    data = {"users": gov_user_id}
-    response = authorized_client.post(url, data)
-    assert response.status_code == 302
-    assert response.url == reverse("queues:cases", kwargs={"queue_pk": queue_pk})
-    assert mock_put_case_case_officer.last_request.json() == {
-        "gov_user_pk": gov_user_id,
-        "case_ids": cases_ids,
-    }
-
-
-def test_case_assignment_case_officer_with_return_to(
-    authorized_client, requests_mock, mock_gov_users, mock_put_case_case_officer
-):
-    cases_ids = [str(uuid.uuid4()), str(uuid.uuid4())]
-    url = (
-        reverse("queues:case_assignments_case_officer", kwargs={"pk": queue_pk})
-        + f"?cases={cases_ids[0]}&cases={cases_ids[1]}&return_to={example_return_to_url}"
-    )
-    data = {"users": gov_user_id}
-    response = authorized_client.post(url, data)
-    assert response.status_code == 302
-    assert response.url == example_return_to_url
-
-
-def test_case_assignment_case_officer_with_invalid_return_to(authorized_client):
-    cases_ids = [str(uuid.uuid4()), str(uuid.uuid4())]
-    url = (
-        reverse("queues:case_assignments_case_officer", kwargs={"pk": queue_pk})
-        + f"?cases={cases_ids[0]}&cases={cases_ids[1]}&return_to=http://www.evil.com"
-    )
-    data = {"users": gov_user_id}
-    response = authorized_client.post(url, data)
-    assert response.status_code == 403
-    assert b"Invalid return_to parameter" in response.content
-
-
-@pytest.mark.parametrize(
-    "user_role_assigned, expected_url_name",
-    (
-        ("CASE_ADVISOR", "case_assignments_assign_user"),
-        ("LU_CASE_OFFICER", "case_assignments_case_officer"),
-    ),
-)
-def test_case_assignment_select_role(authorized_client, mock_gov_user, user_role_assigned, expected_url_name):
-    url_params = (
-        f"?cases={str(uuid.uuid4())}&cases={str(uuid.uuid4())}&return_to={parse.quote(example_return_to_url, safe='')}"
-    )
-
-    url = reverse("queues:case_assignment_select_role", kwargs={"pk": queue_pk}) + url_params
-    response = authorized_client.get(url)
-    assert response.status_code == 200
-
-    response = authorized_client.post(url, {"role": user_role_assigned})
-    assert response.status_code == 302
-    assert response.url == reverse(f"queues:{expected_url_name}", kwargs={"pk": queue_pk}) + url_params
 
 
 def test_tabs_with_all_cases_default(authorized_client, mock_cases_search, mock_cases_search_head):
@@ -741,14 +528,14 @@ def test_queue_assignments(url, authorized_client):
         "ee1a3870-73d7-4af3-b629-e28f2c2227d7": {
             "assignees": [
                 {
-                    "email": "test@mail.com",
+                    "email": "test@mail.com",  # /PS-IGNORE
                     "first_name": "John",
                     "last_name": "Smith",
                     "team_id": "00000000-0000-0000-0000-000000000001",
                     "team_name": "Admin",
                 },
                 {
-                    "email": "test2@mail.com",
+                    "email": "test2@mail.com",  # /PS-IGNORE
                     "first_name": "Joe",
                     "last_name": "Smith",
                     "team_id": "00000000-0000-0000-0000-000000000001",
@@ -819,7 +606,7 @@ def test_activity_updates(url, authorized_client):
     response = authorized_client.get(url)
     expected_activity_updates = [
         {
-            "id": "02cc3048-f893-4f0a-b37f-d066bc0b072a",
+            "id": "02cc3048-f893-4f0a-b37f-d066bc0b072a",  # /PS-IGNORE
             "created_at": "2023-02-02T17:30:05.184293Z",
             "user": {
                 "id": "00000000-0000-0000-0000-000000000001",
@@ -832,10 +619,10 @@ def test_activity_updates(url, authorized_client):
             "additional_text": "additional line1\nadditional line2...",
         },
         {
-            "id": "77d3c3d4-9761-403a-9942-a2fcc41aa35d",
+            "id": "77d3c3d4-9761-403a-9942-a2fcc41aa35d",  # /PS-IGNORE
             "created_at": "2023-02-02T17:30:04.174597Z",
             "user": {
-                "id": "2eb6e0fa-5a5b-4db1-96cc-dd1473e0c636",
+                "id": "2eb6e0fa-5a5b-4db1-96cc-dd1473e0c636",  # /PS-IGNORE
                 "first_name": "Joe",
                 "last_name": "Bloggs",
                 "type": "exporter",
@@ -860,286 +647,10 @@ def test_activity_updates(url, authorized_client):
     assert "applied for a licence." in updates[1]
 
 
-def test_case_assignments_add_user_system_queue(
-    authorized_client,
-    data_queue,
-    data_standard_case,
-    data_assignment,
-    mock_standard_case_with_assignments,
-    mock_queue,
-    mock_gov_users,
-):
-
-    case = data_standard_case
-    url_base = reverse(
-        "queues:case_assignments_assign_user",
-        kwargs={
-            "pk": data_queue["id"],
-        },
-    )
-    url = f"{url_base}?cases={case['case']['id']}"
-    response = authorized_client.get(url)
-    assert response.status_code == 200
-    context = response.context
-    assert isinstance(context["form"], CaseAssignmentUsersForm)
-    # Flow initiated from a system queue, so expect the SELECT_QUEUE step next
-    assert context["wizard"]["steps"].next == CaseAssignmentsCaseAssigneeSteps.SELECT_QUEUE
-
-    html = BeautifulSoup(response.content, "html.parser")
-    assert "Who do you want to allocate as case adviser?" in html.find("h1").get_text()
-
-
-@pytest.fixture
-def mock_users_team_queues_list(requests_mock, gov_uk_user_id, data_queue):
-    url = client._build_absolute_uri(f"/users/{gov_uk_user_id}/team-queues/")
-    return requests_mock.get(url=url, json={"queues": [[data_queue["id"], "Some Queue"]]})
-
-
-@pytest.fixture
-def user_assignment_url(data_standard_case, data_queue):
-    case = data_standard_case
-    url_base = reverse("queues:case_assignments_assign_user", kwargs={"pk": data_queue["id"]})
-    url = f"{url_base}?cases={case['case']['id']}"
-    return url
-
-
-@pytest.fixture
-def post_to_step_user_assignment(post_to_step_factory, user_assignment_url):
-    return post_to_step_factory(user_assignment_url)
-
-
-def test_case_assignments_add_user_system_queue_submit_success(
-    authorized_client,
-    data_queue,
-    data_standard_case,
-    data_assignment,
-    mock_standard_case_with_assignments,
-    mock_queue,
-    mock_gov_users,
-    mock_put_assignments,
-    mock_users_team_queues_list,
-    mock_standard_case_ecju_queries,
-    mock_standard_case_assigned_queues,
-    mock_standard_case_documents,
-    mock_standard_case_additional_contacts,
-    mock_standard_case_activity_filters,
-    post_to_step_user_assignment,
-):
-    data = {
-        "users": [gov_user_id],
-        "note": "foobar",
-    }
-    # POST step 1
-    response = post_to_step_user_assignment(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_USERS,
-        data,
-    )
-    assert response.status_code == 200
-    context = response.context
-    assert isinstance(context["form"], CaseAssignmentQueueForm)
-    assert context["wizard"]["steps"].current == CaseAssignmentsCaseAssigneeSteps.SELECT_QUEUE
-
-    html = BeautifulSoup(response.content, "html.parser")
-    assert "Select a team queue to add the case to" in html.find("h1").get_text()
-
-    data = {
-        "queue": data_queue["id"],
-    }
-    # POST step 2
-    response = post_to_step_user_assignment(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_QUEUE,
-        data,
-        follow=True,
-    )
-    assert response.status_code == 200
-    assert response.redirect_chain[-1][0] == f"/queues/{data_queue['id']}/"
-    messages = [str(msg) for msg in response.context["messages"]]
-    expected_message = "Case adviser was added successfully"
-    assert messages == [expected_message]
-    assert mock_put_assignments.called
-    assert mock_put_assignments.last_request.json() == {
-        "case_assignments": [
-            {
-                "case_id": data_standard_case["case"]["id"],
-                "users": [gov_user_id],
-            }
-        ],
-        "note": "foobar",
-        "remove_existing_assignments": False,
-    }
-
-    html = BeautifulSoup(response.content, "html.parser")
-    assert expected_message in html.select("div.app-snackbar__content")[0].get_text()
-
-
-def test_case_assignments_add_user_system_queue_submit_validation_error(
-    authorized_client,
-    data_queue,
-    data_standard_case,
-    data_assignment,
-    mock_standard_case_with_assignments,
-    mock_queue,
-    mock_gov_users,
-    mock_put_assignments,
-    mock_users_team_queues_list,
-    mock_standard_case_ecju_queries,
-    mock_standard_case_assigned_queues,
-    mock_standard_case_documents,
-    mock_standard_case_additional_contacts,
-    mock_standard_case_activity_filters,
-    post_to_step_user_assignment,
-):
-    data = {
-        "note": "foobar",
-    }
-    # POST step 1 - validation error
-    response = post_to_step_user_assignment(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_USERS,
-        data,
-    )
-    assert response.status_code == 200
-    assert response.context["form"]["users"].errors == ["Select a user to allocate"]
-
-    data = {
-        "users": [gov_user_id],
-        "note": "foobar",
-    }
-    # POST step 1 - valid
-    response = post_to_step_user_assignment(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_USERS,
-        data,
-    )
-    assert response.status_code == 200
-    assert isinstance(response.context["form"], CaseAssignmentQueueForm)
-    assert response.context["wizard"]["steps"].current == CaseAssignmentsCaseAssigneeSteps.SELECT_QUEUE
-
-    html = BeautifulSoup(response.content, "html.parser")
-    assert "Select a team queue to add the case to" in html.find("h1").get_text()
-
-    data = {}
-    # POST step 2
-    response = post_to_step_user_assignment(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_QUEUE,
-        data,
-    )
-    assert response.status_code == 200
-    assert response.context["form"]["queue"].errors == ["Select a queue to add the case to"]
-
-
-@pytest.fixture
-def mock_team_queue(requests_mock, data_queue):
-    data_queue["is_system_queue"] = False
-    url = client._build_absolute_uri("/queues/")
-    yield requests_mock.get(url=re.compile(f"{url}.*/"), json=data_queue)
-
-
-@pytest.fixture
-def mock_put_assignments(requests_mock, data_queue, data_assignment):
-    url = client._build_absolute_uri(f"/queues/{data_queue['id']}/case-assignments/")
-    return requests_mock.put(url=url, json=data_assignment)
-
-
-def test_case_assignments_add_user_team_queue(
-    authorized_client,
-    data_queue,
-    data_standard_case,
-    data_assignment,
-    mock_standard_case_with_assignments,
-    mock_team_queue,
-    mock_gov_users,
-):
-
-    case = data_standard_case
-    url_base = reverse("queues:case_assignments_assign_user", kwargs={"pk": data_queue["id"]})
-    url = f"{url_base}?cases={case['case']['id']}"
-    response = authorized_client.get(url)
-    assert response.status_code == 200
-    context = response.context
-    assert isinstance(context["form"], CaseAssignmentUsersForm)
-    # Flow initiated from a team queue, so expect no followup step
-    assert context["wizard"]["steps"].next == None
-
-    html = BeautifulSoup(response.content, "html.parser")
-    assert "Who do you want to allocate as case adviser?" in html.find("h1").get_text()
-
-
-def test_case_assignments_add_user_team_queue_submit_success(
-    authorized_client,
-    data_queue,
-    data_standard_case,
-    data_assignment,
-    mock_team_queue,
-    mock_gov_users,
-    mock_put_assignments,
-    mock_standard_case_with_assignments,
-    mock_standard_case_ecju_queries,
-    mock_standard_case_assigned_queues,
-    mock_standard_case_documents,
-    mock_standard_case_additional_contacts,
-    mock_standard_case_activity_filters,
-    post_to_step_user_assignment,
-):
-    data = {
-        "users": [gov_user_id],
-        "note": "foobar",
-    }
-    response = post_to_step_user_assignment(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_USERS,
-        data,
-        follow=True,
-    )
-    assert response.status_code == 200
-    assert response.redirect_chain[-1][0] == f"/queues/{data_queue['id']}/"
-    messages = [str(msg) for msg in response.context["messages"]]
-    expected_message = "Case adviser was added successfully"
-    assert messages == [expected_message]
-    assert mock_put_assignments.called
-    assert mock_put_assignments.last_request.json() == {
-        "case_assignments": [
-            {
-                "case_id": data_standard_case["case"]["id"],
-                "users": [gov_user_id],
-            }
-        ],
-        "note": "foobar",
-        "remove_existing_assignments": False,
-    }
-
-    html = BeautifulSoup(response.content, "html.parser")
-    assert expected_message in html.select("div.app-snackbar__content")[0].get_text()
-
-
-def test_case_assignments_add_user_team_queue_submit_validation_error(
-    authorized_client,
-    data_queue,
-    data_standard_case,
-    data_assignment,
-    mock_team_queue,
-    mock_gov_users,
-    mock_put_assignments,
-    mock_standard_case_with_assignments,
-    mock_standard_case_ecju_queries,
-    mock_standard_case_assigned_queues,
-    mock_standard_case_documents,
-    mock_standard_case_additional_contacts,
-    mock_standard_case_activity_filters,
-    post_to_step_user_assignment,
-):
-
-    data = {
-        "users": [],
-        "note": "foobar",
-    }
-    response = post_to_step_user_assignment(
-        CaseAssignmentsCaseAssigneeSteps.SELECT_USERS,
-        data,
-    )
-    assert response.status_code == 200
-    assert response.context["form"]["users"].errors == ["Select a user to allocate"]
-
-
 def test_filter_none_pending_gov_users(authorized_client, mock_cases_search):
     url = reverse("queues:cases")
     response = authorized_client.get(url)
     gov_users = response.context["data"]["results"]["filters"]["gov_users"]
-    assert gov_users == [{"full_name": "John Smith", "id": "2a43805b-c082-47e7-9188-c8b3e1a83cb0", "pending": False}]
+    assert gov_users == [
+        {"full_name": "John Smith", "id": "2a43805b-c082-47e7-9188-c8b3e1a83cb0", "pending": False}
+    ]  # /PS-IGNORE
